# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.DeleteAccountRequest:
    properties:
      confirmation:
        type: string
      password:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.UpdateProfileRequest:
    properties:
      display_name:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_adapter_http_dto.UserProfileResponse:
    properties:
      auth_methods:
        items:
          type: string
        type: array
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      is_active:
        type: boolean
      stats:
        $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.UserProfileStats'
      updated_at:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_adapter_http_dto.UserProfileStats:
    properties:
      app_password_count:
        type: integer
      calendar_count:
        type: integer
      contact_count:
        type: integer
    type: object
...
  internal_adapter_http.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: error_code
        type: string
      message:
        example: Human readable message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HTTP Basic Authentication for DAV endpoints
  title: CalDAV/CardDAV Server API
  version: "1.0"
paths:
...
  /users/me:
    delete:
      consumes:
      - application/json
      description: Permanently delete user account
      parameters:
      - description: Delete confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.DeleteAccountRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - Users
    get:
      description: Get current user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update current user's display name
      parameters:
      - description: Profile updates
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Users
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...